Stable Fixes + Enhancement

+++ commands/poll.js
@@ -48,6 +48,7 @@ module.exports = {
           error: ["error"]
         })
         .catch(err => msg.edit("Time's up. Cancelled action."));
+      if(!channelCollected.first()) return msg.edit("Time's up. Cancelled action.");
       if (channelCollected.first().content === "cancel") {
         await channelCollected.first().delete();
         return msg.edit("Cancelled poll.");
@@ -69,11 +70,12 @@ module.exports = {
       );
       var collected = await message.channel
         .awaitMessages(filter, {
-          time: 30000,
+          time: 60000,
           max: 1,
           error: ["time"]
         })
         .catch(err => msg.edit("Time's up. Cancelled action."));
+      if(!collected.first()) return msg.edit("Time's up. Cancelled action.");
       if (collected.first().content === "cancel") {
         await collected.first().delete();
         return msg.edit("Cancelled poll.");
@@ -92,7 +94,7 @@ module.exports = {
           error: ["time"]
         })
         .catch(err => msg.edit("Time's up. Cancelled action."));
-      if (collected2.first() === undefined)
+      if (!collected2.first())
         msg.edit("Time's up. Cancelled action.");
       if (collected2.first().content === "cancel") {
         await collected2.first().delete();
@@ -151,6 +153,7 @@ module.exports = {
           error: ["time"]
         })
         .catch(err => msg.edit("Time's up. Cancelled action."));
+      if(!optionString.first()) return msg.edit("Time's up. Cancelled action.");
       if (optionString.first().content === "cancel") {
         await optionString.first().delete();
         return msg.edit("Cancelled poll.");
